Ultima actualizacion por que dejaron sin electricidad el pueblo

+++ demo_cliente/src/main/java/count/proyecto/web/demo_cliente/Models/Compra.java
@@ -19,7 +19,7 @@ public class Compra {
     private long idCompra;
 
     @Column(name = "totalCompra")
-    private long totalCompra;
+    private double totalCompra;
 
     @Column (name="fecha")
     private Date fecha;
